requirements.txt
# Currently used dependencies
# ML libs
transformers>=4.48.0
scipy>=1.10.0
numpy>=1.24.0
torch>=2.4.0

# audio libs
suno-bark>=0.1.5

# utils
tqdm>=4.67.1
python-dotenv>=0.19.0
requests>=2.32.3
httpx>=0.28.1
PyPDF2>=3.0.0
accelerate>=0.26.0
bitsandbytes>=0.42.0

# LLMs and chains
openai>=1.59.8
llamaapi>=0.1.36
llama-parse>=0.5.19
# langchain>=0.3.14
# langchain-community>=0.0.10 # DDoS CVE
# langchain-huggingface>=0.1.2

# API and serving
fastapi>=0.115.6
pydantic>=2.10.5
uvicorn>=0.34.0
python-multipart>=0.0.20
aiohttp>=3.11.11
backoff>=2.2.1

# Additional utils
ipython>=8.31.0
optimum>=1.23.3
pydub>=0.25.1
python-ffmpeg>=2.0.12

# New dependency
aiofiles>=23.2.1
beautifulsoup4>=4.12.2


# Cerebras dependencies
cerebras_cloud_sdk>=1.16.0
groq>=0.18.0
#cerebras_pytorch>=2.4.0
optillm>=0.1.4
